Global Styles: refactor site background controls and move site global styles into Background group (#65304)
Expose background styles in the top level global styles navigation menu. Background is now underneath color. Refactor background image controls and their styles into a dedicated component. This is in preparation for adding colors and other background controls later.
