+++ packages/block-editor/src/style.scss
@@ -1,4 +1,5 @@
 @import "./autocompleters/style.scss";
+@import "./components/background-image-control/style.scss";
 @import "./components/block-alignment-control/style.scss";
 @import "./components/block-canvas/style.scss";
 @import "./components/block-icon/style.scss";
